requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
dateutils==0.6.6
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.3
django-toolbelt==0.0.1
future==0.16.0
google-api-python-client==1.6.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
Jinja2==2.9.6
line-bot-sdk==1.3.0
MarkupSafe==1.0
oauth2client==4.1.2
psycopg2==2.7.1
py==1.4.34
pyasn1==0.2.3
pyasn1-modules==0.0.9
Pygments==2.2.0
pytest==3.2.1
python-dateutil==2.6.0
python-dotenv==0.7.1
pytz==2017.2
requests==2.18.1
rsa==3.4.2
selenium==3.4.3
six==1.10.0
static3==0.7.0
uritemplate==3.0.0
urllib3==1.21.1
